polish TabbedPage & TabbedPageTab
