+++ app/components/layout/TabbedPage/TabbedPageTab.jsx
@@ -1,0 +1,8 @@
+const TabbedPageTab = ({ children }) => children;
+
+TabbedPageTab.propTypes = {
+  path: PropTypes.string.isRequired,
+  link: PropTypes.node.isRequired
+};
+
+export default TabbedPageTab;
