+++ app/components/layout/TabbedPage/index.js
@@ -1,0 +1,2 @@
+export { default as TabbedPageTab } from "./TabbedPageTab";
+export { default as TabbedPage } from "./TabbedPage";
