+++ app/components/layout/index.js
@@ -1,4 +1,4 @@
-export { default as TabbedPage } from "./TabbedPage";
+export { TabbedPage } from "./TabbedPage";
 export { TabbedPageTab } from "./TabbedPage";
 export { default as StandaloneHeader } from "./StandaloneHeader";
 export * from "./TitleHeader";
